# app/docs/swagger.yaml
consumes:
- application/json
definitions:
  apperrors.ErrorCode:
    enum:
    - access_token_expired
    - all_signals_failed_processing
    - authentication_error
    - authorization_error
    - client_closed
    - database_error
    - forbidden
    - invalid_correlation_id
    - internal_error
    - invalid_request
    - invalid_url_param
    - malformed_body
    - not_implemented
    - network_error
    - password_too_short
    - refresh_token_invalid
    - request_too_large
    - unsupported_media_type
    - rate_limit_exceeded
    - resource_already_exists
    - resource_expired
    - resource_in_use
    - resource_not_found
    - timeout
    - token_creation_failed
    type: string
    x-enum-varnames:
    - ErrCodeAccessTokenExpired
    - ErrCodeAllSignalsFailedProcessing
    - ErrCodeAuthenticationFailure
    - ErrCodeAuthorizationFailure
    - ErrCodeClientClosed
    - ErrCodeDatabaseError
    - ErrCodeForbidden
    - ErrCodeInvalidCorrelationID
    - ErrCodeInternalError
    - ErrCodeInvalidRequest
    - ErrCodeInvalidURLParam
    - ErrCodeMalformedBody
    - ErrCodeNotImplemented
    - ErrCodeNetworkError
    - ErrCodePasswordTooShort
    - ErrCodeRefreshTokenInvalid
    - ErrCodeRequestTooLarge
    - ErrCodeUnsupportedMediaType
    - ErrCodeRateLimitExceeded
    - ErrCodeResourceAlreadyExists
    - ErrCodeResourceExpired
    - ErrCodeResourceInUse
    - ErrCodeResourceNotFound
    - ErrCodeTimeout
    - ErrCodeFailedToCreateToken
  auth.AccessTokenResponse:
    properties:
      access_token:
        description: AccessToken is a JWT access token containing claims about the
          account and its permissions (see Claims struct)
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJTaWduYWxzZCIsInN1YiI6ImQwODkzODllLTg0Y2MtNDU4MC1hNDBjLTNmMGEyMGU5YWFmMSIsImV4cCI6MTc3NDg1OTg1MywiaWF0IjoxNzc0ODU4MDUzLCJhY2NvdW50X2lkIjoiZDA4OTM4OWUtODRjYy00NTgwLWE0MGMtM2YwYTIwZTlhYWYxIiwiYWNjb3VudF90eXBlIjoidXNlciIsInJvbGUiOiJzaXRlYWRtaW4iLCJpc25fcGVybXMiOnsiYWRtaW4taXNuIjp7ImNhbl9yZWFkIjp0cnVlLCJjYW5fd3JpdGUiOnRydWUsImNhbl9hZG1pbmlzdGVyIjp0cnVlLCJzaWduYWxfdHlwZXMiOnsiYWRtaW4taXNuLXNpZ25hbC92MS4wLjAiOnsiaW5fdXNlIjp0cnVlfX0sInZpc2liaWxpdHkiOiJwcml2YXRlIiwiaW5fdXNlIjpmYWxzZX0sInNpdGVhZG1pbi1pc24iOnsiY2FuX3JlYWQiOnRydWUsImNhbl93cml0ZSI6dHJ1ZSwiY2FuX2FkbWluaXN0ZXIiOnRydWUsInNpZ25hbF90eXBlcyI6eyJzaXRlYWRtaW4taXNuLXNpZ25hbC92MS4wLjAiOnsiaW5fdXNlIjp0cnVlfX0sInZpc2liaWxpdHkiOiJwcml2YXRlIiwiaW5fdXNlIjp0cnVlfX19.03SEhY_tL2dbNjMYTHka_-5DhR5FhNh0CI0etsQiEo8
        type: string
      account_id:
        description: AccountID is the account id of the user making the request
        example: a38c99ed-c75c-4a4a-a901-c9485cf93cf3
        type: string
      account_type:
        description: AccountType is the account type of the user making the request
          (user or service_account)
        enum:
        - user
        - service_account
        type: string
      email:
        description: email is the contact email for the account
        type: string
      expires_in:
        description: ExpiresIn is the token expiry in seconds
        example: 1800
        type: integer
      isn_perms:
        additionalProperties:
          $ref: '#/definitions/auth.IsnPerm'
        description: IsnPerms is a map of the ISNs the account has access to and the
          permissions granted (the map key is the isn_slug)
        type: object
      role:
        description: Role is the role of the user making the request (siteadmin, isnadmin,
          member)
        enum:
        - siteadmin
        - isnadmin
        - member
        example: isnadmin
        type: string
      token_type:
        description: TokenType (Bearer) - used as a prompt for the client to use the
          Bearer token type when making requests
        example: Bearer
        type: string
    type: object
  auth.IsnPerm:
    properties:
      can_administer:
        description: CanAdminister is true if the account is the owner of the isn
          or a site admin
        example: false
        type: boolean
      can_read:
        description: CanRead is true if the account has read access to the isn
        example: true
        type: boolean
      can_write:
        description: CanWrite is true if the account has write access to the isn
        example: false
        type: boolean
      in_use:
        description: InUse is true if the isn is active
        example: true
        type: boolean
      signal_types:
        additionalProperties:
          $ref: '#/definitions/auth.SignalType'
        description: SignalTypes is a map of the signal type paths to the signal type
          details (key is the signal type path)
        type: object
      visibility:
        description: Visibility is the ISN visibility setting (public or private)
        enum:
        - public
        - private
        example: private
        type: string
    type: object
  auth.SignalType:
    properties:
      in_use:
        description: InUse is true if the signal type is active. Note this will always
          be false if the parent ISN has been marked as 'not in use'
        example: true
        type: boolean
      path:
        description: |-
          Path is the signal type path in the format "slug/v{version}"
          This is the key used in the SignalTypes map in IsnPerms
        example: sample-signal/v0.0.1
        type: string
      readme_url:
        description: ReadmeURL is the URL of the signal type readme
        example: https://github.com/user/project/blob/2025.01.01/readme.md
        type: string
      schema_url:
        description: SchemaURL is the URL of the signal type schema
        example: https://github.com/user/project/blob/2025.01.01/schema.json
        type: string
      sem_ver:
        description: SemVer is the signal type version (e.g. 0.0.1)
        example: 0.0.1
        type: string
      slug:
        description: Slug is the signal type slug (unique per site)
        example: sample-signal
        type: string
    type: object
...
  handlers.IsnAccount:
    properties:
      account_id:
        example: a38c99ed-c75c-4a4a-a901-c9485cf93cf3
        type: string
      account_role:
        enum:
        - siteadmin
        - isnadmin
        - member
        example: isnadmin
        type: string
      account_type:
        enum:
        - user
        - service_account
        example: user
        type: string
      can_read:
        example: true
        type: boolean
      can_write:
        example: false
        type: boolean
      client_id:
        example: sa_exampleorg_k7j2m9x1
        type: string
      client_organization:
        example: Example Organization
        type: string
      created_at:
        example: "2025-06-03T13:47:47.331787+01:00"
        type: string
      email:
        example: user@example.com
        type: string
      id:
        example: 67890684-3b14-42cf-b785-df28ce570400
        type: string
      is_active:
        example: true
        type: boolean
      isn_id:
        example: 67890684-3b14-42cf-b785-df28ce570400
        type: string
      updated_at:
        example: "2025-06-03T13:47:47.331787+01:00"
        type: string
    type: object
...
  handlers.LoginRequest:
    properties:
      email:
        example: example@example.com
        type: string
      password:
        description: passwords must be at least 11 characters long
        example: lkIB53@6O^Y
        type: string
    type: object
...
  handlers.UpdateIsnAccountPermissionRequest:
    properties:
      can_read:
        example: true
        type: boolean
      can_write:
        example: false
        type: boolean
    type: object
...
  responses.ErrorResponse:
    properties:
      error_code:
        allOf:
        - $ref: '#/definitions/apperrors.ErrorCode'
        description: Standard app error code
        example: error_code
      message:
        description: Client message
        example: client message
        type: string
      request_id:
        description: request tracking id
        example: a1b2c3d4
        type: string
    type: object
...
info:
  contact: {}
  description: |-
    Signals ISN service API for managing Information Sharing Networks

    ## Common Error Responses
    All endpoints may return:
    - `400` Malformed request (invalid json, missing required fields, etc.)
    - `401` Unauthorized (invalid credentials)
    - `403` Forbidden (insufficient permissions)
    - `404` Resource not found (e.g. an ISN or signal type that doesn't exist or is not in use)
    - `413` Request body exceeds size limit
    - `429` Rate limit exceeded
    - `500` Internal server error
    - `504` Request timed out (`timeout`)

    Individual endpoints document their specific business logic errors.

    Standard error responses include a JSON response body containing `error_code`, `message` and `request_id` fields.

    The /oauth endpoints return `error` (the RFC6749 standard error code), `error_code` and `error_description` fields instead.

    ## Request Limits
    The endpoints are protected by:
    - **Rate limiting**: Configurable requests per second (all endpoints)
    - **Request size limits** (configurable - these are the defaults): 64KB for admin/auth endpoints, 5MB for json and event signal submissions, 20MB for document uploads

    Check the `Signalsd-Max-Request-Size` response header for the configured size limit of the endpoint called.

    The rate limit is set globally and prevents abuse of the service.
...

    ## Authentication & Authorization

    ### OAuth
...

...

...
    ```
...
    ```

...
paths:
...
  /api/auth/login:
    post:
      description: |-
        The response body includes an access token which can be used to access the protected endpoints, assuming the account has the appropriate permissions.
        The access_token is valid for 30 minutes.

        As part of the login response, the server sets a http-only cookie on the client that will allow it to refresh the token (POST to the /oauth/token endpoint with grant_type=refresh_token in the form body).
        The cookie's path is `/oauth`, so browsers only send it to the /oauth endpoints.
        The refresh_token lasts 30 days unless it is revoked earlier.
        - Each refresh issues a new refresh_token valid for another 30 days, so the user only has to log in again if they don't refresh for 30 days.
        - Logging in revokes the user's other refresh tokens, so a user has one active session at a time.
        - To revoke the refresh_token, call the /oauth/revoke endpoint.

        The account's role and permissions are encoded as part of the jwt access token and this information is also provided in the response body.
      parameters:
      - description: email and password
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.LoginRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.AccessTokenResponse'
        "400":
          description: malformed_body
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "401":
          description: authentication_error
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "500":
          description: database_error | token_creation_failed
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      summary: Login
      tags:
      - User Authentication
...
  /api/isn/{isn_slug}/accounts:
    get:
      description: |-
        Get a list of all accounts (users and service accounts) that have had permissions on the specified ISN.
        Accounts whose access has been revoked are included (with can_read and can_write both false).
        Only ISN admins (for the ISNs they own) and site admins can view this information
      parameters:
      - description: ISN slug
        example: sample-isn
        in: path
        name: isn_slug
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handlers.IsnAccount'
            type: array
        "401":
          description: authorization_error | access_token_expired
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "403":
          description: forbidden
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "404":
          description: resource_not_found
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "500":
          description: database_error
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      security:
      - BearerAccessToken: []
      summary: Get ISN Account Membership
      tags:
      - ISN Configuration
  /api/isn/{isn_slug}/accounts/{account_id}:
    put:
      description: |-
        Update an account's access permission for an ISN. Set both can_read and can_write to false to revoke all access.

        This endpoint can only be used by admin accounts:
        - ISN admins can only update membership for ISNs they own.
        - Site admins can update membership for any ISN

        Permissions:
        - Accounts with 'read' permission can view all signals on the ISN.
        - Accounts with 'write' permission can create signals on the ISN.
        - For accounts that need read/write access to an ISN, you must grant both 'read' and 'write' permissions.

        Note that accounts with only 'write' permission to an ISN can view the signals they created, and the signals other accounts have correlated to their signals, but no others.

        The change takes effect when the account next gets an access token (up to 30 minutes later).

        You must supply values for both can_read and can_write.
      parameters:
      - description: permission details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateIsnAccountPermissionRequest'
      - description: ISN slug
        example: sample-isn
        in: path
        name: isn_slug
        required: true
        type: string
      - description: account id
        example: a38c99ed-c75c-4a4a-a901-c9485cf93cf3
        in: path
        name: account_id
        required: true
        type: string
      responses:
        "200":
          description: OK
        "400":
          description: invalid_request | malformed_body
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "401":
          description: authorization_error | access_token_expired
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "403":
          description: forbidden
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "404":
          description: resource_not_found
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "500":
          description: database_error
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      security:
      - BearerAccessToken: []
      summary: Grant/Revoke ISN Access
      tags:
      - Account Management
...
produces:
...
securityDefinitions:
  BearerAccessToken:
    description: Bearer {JWT access token}
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...